requirements.txt
# Core
django==4.2.9
python-dotenv
django-environ
str2bool
Pillow
djangorestframework
markdown
django-filter
factory_boy
django-autoslug
django-countries
django-phonenumber-field
phonenumbers

# UI
django-admin-argon-dashboard==1.0.16
fontawesomefree

# Deployment
whitenoise
gunicorn
uvicorn

psycopg2-binary
dj-database-url